Move base submodule to 'private' namespace
